+++ src/CompileOptions.php
@@ -8,10 +8,16 @@ final class CompileOptions
     private $namespaces = [];
 
     /** @var string */
+    private $domain;
+
+    /** @var string */
     private $output;
 
     /** @var string */
     private $manifest;
+
+    /** @var bool */
+    private $include_all;
 
     /** @var \Closure */
     private $callback;
@@ -41,6 +47,14 @@ final class CompileOptions
     /**
      * @return string
      */
+    public function getDomain()
+    {
+        return $this->domain;
+    }
+
+    /**
+     * @return string
+     */
     public function getOutput()
     {
         return $this->output;
@@ -52,6 +66,14 @@ final class CompileOptions
     public function getManifest()
     {
         return $this->manifest;
+    }
+
+    /**
+     * @return bool
+     */
+    public function getIncludeAll()
+    {
+        return $this->include_all;
     }
 
     /**