+++ src/Validator/SchemaMustContainsField.php
@@ -13,8 +13,8 @@ class SchemaMustContainsField implements Constraint
     public function validate(SchemaDescriptor $a, SchemaDescriptor $b)
     {
         $diff = array_diff(
-            array_keys(array_merge($a->getInheritedFields(), $a->getFields())),
-            array_keys(array_merge($b->getInheritedFields(), $b->getFields()))
+            array_keys(array_merge($a->getFields(), $a->getInheritedFields())),
+            array_keys(array_merge($b->getFields(), $b->getInheritedFields()))
         );
         if (count($diff)) {
             throw new ValidatorException(sprintf(